#330 Moved genomes-related functions from utils and models to portal.config.genomes.
